requirements.txt
requests==2.23.0
ecdsa==0.17.0
pycryptodome==3.9.8
crcmod==1.7
eth_utils>=0.9.0
